Skip Twisted test instad of not declaring them
It's easy to miss that there are tests for running against Twisted. I
hope to make it a bit more obvious by raising a SkipTest during test
collection, and extracting them into their own module.
